video-detector-inference/requirements.txt
huggingface_hub
gradio
numpy==1.21.5
pandas==1.4.2
Pillow==9.0.1
#dlib==19.24.0
imageio==2.9.0
imgaug==0.4.0
tqdm==4.61.0
scipy==1.7.3
seaborn==0.11.2
pyyaml==6.0
imutils==0.5.4
opencv-python==4.6.0.66
scikit-image==0.19.2
scikit-learn==1.0.2
albumentations==1.1.0
torch==1.12.0
torchvision==0.13.0
torchaudio==0.12.0
efficientnet-pytorch==0.7.1
timm==0.6.12
segmentation-models-pytorch==0.3.2
torchtoolbox==0.1.8.2
tensorboard==2.10.1
setuptools==59.5.0
loralib
einops
transformers
filterpy
simplejson
kornia
fvcore
imgaug==0.4.0
git+https://github.com/openai/CLIP.git